feat: Service, Repository 구현 #8

+++ src/main/java/umc/study/web/dto/ReviewResponseDTO.java
@@ -19,7 +19,7 @@ public class ReviewResponseDTO {
 
         Long reviewId;
 
-        Integer rating;
+        double rating;
 
         String content;
 
@@ -47,7 +47,7 @@ public class ReviewResponseDTO {
     @AllArgsConstructor
     public static class ReviewPreViewDTO {
         String ownerNickname;
-        Float score;
+        double score;
         String body;
         LocalDate createdAt;
     }
